+++ generators/python/src/fern_python/snippet/snippet_endpoint_expression.py
@@ -1,5 +1,5 @@
-from typing import Optional
 from dataclasses import dataclass
+from typing import Optional
 
 import fern.generator_exec.resources as generator_exec