+++ seed/ruby-model/enum/lib/types_export.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "seed_enum_client/types/operand"
+require_relative "seed_enum_client/types/color"
+require_relative "seed_enum_client/types/color_or_operand"