+++ seed/ruby-model/error-property/lib/types_export.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "seed_error_property_client/errors/types/property_based_error_test_body"