+++ seed/ruby-model/imdb/lib/types_export.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "seed_api_client/imdb/types/movie"
+require_relative "seed_api_client/imdb/types/create_movie_request"