+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.19.1] - 2024-05-23
+
+- Internal: SDK snippets now specify their example ID and we create a snippet per example
+
 ## [0.19.0] - 2024-05-20
 
 - Feature: Add `inlineFileProperties` configuration to support generating file upload properties